+++ packages/terra-form-select/CHANGELOG.md
@@ -8,10 +8,8 @@ Unreleased
 * Updated translations.
 
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * updated package.json test scripts
-
-
-### Changed
 * Update tests for dev-site v6
 
 5.25.0 - (July 30, 2019)
@@ -46,6 +44,7 @@ Unreleased
 5.20.0  - (July 11, 2019)
 ------------------
 ### Changed
+* Components updated to use `injectIntl` to interface with `react-intl's` `intl` context.
 * Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
 * Updated Docs around the requirement of terra-base and aggregate-translations.